refactor: create EGSwitchOrgModal for handling org switching (addresses EG-1096 and EG-1098)
